Refactor selection responsability
